refactor: Move `Digest` into `source.rs`
It makes more sense for `Source` to be responsible for gettings its
own `Digest`.

Impl this as an inherent method so types which impl `AsRef<Source>`
can easily call this.
